Fix test_pkey_dsa.rb in FIPS.
Note that we created the `dsa2048.pem` and signature text
(`signature_encoded.txt`) by the following steps.

```
$ export OPENSSL_CONF=/path/to/ssl/openssl_fips.cnf

$ /path/to/bin/openssl dsaparam -out dsaparam2048.pem 2048
$ /path/to/bin/openssl gendsa -out dsa2048.pem dsaparam2048.pem

$ echo -n "Sign me!" > data.txt
$ /path/to/bin/openssl dgst -sha256 -sign dsa2048.pem data.txt > signature.txt
$ cat signature.txt | base64 > signature_encoded.txt
```

Skip the `test_DSAPrivateKey_encrypted` on FIPS because AES-128-CBC, the
password based encryption used in the PEM format uses MD5 for deriving the
encryption key from the password, and MD5 is not FIPS-approved.
See also the comment on the `test/openssl/utils.rb#omit_on_fips`.