test/openssl/test_pkey_dsa.rb: skip all tests in FIPS mode
It appears that the next version of OpenSSL running in the FIPS mode
will not support DSA key generation and signature generation. Although
it appears that FIPS 186-5 still allows signature verification, there
seems to be little point in writing such a specific test case.

DSA will still be tested on OpenSSL master if it is not operating in
the FIPS mode.

test_generate_on_non_fips is merged to test_generate again.